--- a/include/crow/http_server.h
+++ b/include/crow/http_server.h
@@ -50,7 +50,7 @@ namespace crow
         {
             tick_function_();
             tick_timer_.expires_after(std::chrono::milliseconds(tick_interval_.count()));
-            tick_timer_.async_wait([this](const std::error_code& ec) {
+            tick_timer_.async_wait([this](const asio::error_code& ec) {
                 if (ec)
                     return;
                 on_tick();
@@ -126,7 +126,7 @@ namespace crow
             {
                 tick_timer_.expires_after(std::chrono::milliseconds(tick_interval_.count()));
                 tick_timer_.async_wait(
-                  [this](const std::error_code& ec) {
+                  [this](const asio::error_code& ec) {
                       if (ec)
                           return;
                       on_tick();
@@ -141,7 +141,7 @@ namespace crow
             CROW_LOG_INFO << "Call `app.loglevel(crow::LogLevel::Warning)` to hide Info level logs.";
 
             signals_.async_wait(
-              [&](const std::error_code& /*error*/, int /*signal_number*/) {
+              [&](const asio::error_code& /*error*/, int /*signal_number*/) {
                   stop();
               });
 
@@ -225,7 +225,7 @@ namespace crow
 
                 acceptor_.async_accept(
                   p->socket(),
-                  [this, p, &is, service_idx](std::error_code ec) {
+                  [this, p, &is, service_idx](asio::error_code ec) {
                       if (!ec)
                       {
                           is.post(