--- a/include/crow/http_server.h
+++ b/include/crow/http_server.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <chrono>
-#include <boost/date_time/posix_time/posix_time.hpp>
 #define ASIO_STANDALONE
 #include <asio.hpp>
 #ifdef CROW_ENABLE_SSL