--- a/include/crow/json.h
+++ b/include/crow/json.h
@@ -1898,7 +1898,7 @@ namespace crow
 
 
 
-        //std::vector<boost::asio::const_buffer> dump_ref(wvalue& v)
+        //std::vector<asio::const_buffer> dump_ref(wvalue& v)
         //{
         //}
     } // namespace json