--- a/include/crow/parser.h
+++ b/include/crow/parser.h
@@ -2,7 +2,6 @@
 
 #include <string>
 #include <unordered_map>
-#include <boost/algorithm/string.hpp>
 #include <algorithm>
 
 #include "crow/http_parser_merged.h"