This Pr removes `Boost` as a dependency and will instead rely on the standalone `asio` library.
CI now installs `libasio-dev` instead of `libboost-all-dev` and the documentation resembles that change. (Although the MacOS section could use a double-check, I have no such device on hand)

### Proper testing necessary
Unittests are all passing but I think some real-life testing does no harm.

### Removed Boost functions and types and their replacements:
- `boost::lexical_cast` becomes `crow::utility::lexical_cast`. Its implementation relies on `std::stringstream` for converting between types.
- `boost::hash_combine` is replaced by a private function. Its implementation idea stems from around the internet and can be found in multiple sources.
- `boost::iequals` is replaced by `crow::utility::string_equals` taking two mandatory `std::string` arguments and one optional `bool` controlling the case-sensitivity.
- `boost::system::error_code` is replaced by `asio::error_code` which is an alias for `std::error_code`.
- Every `boost::asio::` function is replaced with their standalone `asio::` counterpart.
- `boost::array` is replaced with `std::array`.
- `boost::posix_time` is replaced with a `std::chrono` equivalent.
- `boost::*_comparable` is dropped and the respective operators are implemented manually.
- `boost::lexicographical_compare` is replaced with `std::lexicographical_compare`.
- `boost::optional` is replaced with `std::unique_ptr`.
- `boost::trim` is replaced with `crow::utility::trim`.

### Introduced functions (\*explicitly\* tested?):
- `private void crow::ci_hash::hash_combine`. (❌)
- `private bool crow::json::detail::r_string::equals`. (❌)
- Copy-constructor for `crow::CookieParser::Cookie`. (❌)
- `bool crow::utility::string_equals` (✅)
- `T crow::utility::lexical_cast` (✅)
- `std::string crow::utility::trim` (✅)