@luca-schlecker 2022-06-07
I'll have to look into that unittest again. Passing fine on my system.
Unfortunately, I had to make a change to `query_string.h` and in the process vscode auto-formatted everything according to Crow's guidelines. The `yml` files were also formatted on save, reversed that change though.