@The-EDev 2022-06-09
```sh
$ g++-4.8 example_with_all.cpp -std=c++11 -lpthread -o examp
In file included from example_with_all.cpp:1:0:
crow_all.h: In member function ‘std::string crow::CookieParser::Cookie::dump() const’:
crow_all.h:5537:27: error: ‘put_time’ is not a member of ‘std’
                        << std::put_time(expires_at_.get(), HTTP_DATE_FORMAT);
                           ^
```

it looks like the code doesn't compile with GCC4.8's c++11, which is either a bug in GCC or a missing feature from c++11