@luca-schlecker 2022-06-09
This PR doesn't add the `std::put_time` call though, it was there before.