@The-EDev 2022-06-11
So here's what I did, I modified the `helloworld` example to run multithreaded and then again to run no logs. Then I ran `wrk` with 8 connections and 8 threads on both a `master` build and a boostless one. The results may not be accurate as to Crow's actual performance, but they are useful in measuring the difference between boostless and boostful Crow:
```
-- WITH LOG --

- NoBoost -
Running 1m test @ http://0.0.0.0:18080/
  8 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   183.56us  177.83us  16.45ms   95.69%
    Req/Sec     5.75k   188.66     6.88k    75.31%
  2750033 requests in 1.00m, 351.43MB read
Requests/sec:  45757.82
Transfer/sec:      5.85MB

- Master -
Running 1m test @ http://0.0.0.0:18080/
  8 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   655.24us  166.19us  12.31ms   80.37%
    Req/Sec     1.53k   276.82     8.51k    70.84%
  730762 requests in 1.00m, 93.39MB read
Requests/sec:  12159.17
Transfer/sec:      1.55MB


-- NO LOG --

- NoBoost -
Running 1m test @ http://0.0.0.0:18080/
  8 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    58.25us  276.31us  23.32ms   99.38%
    Req/Sec    21.20k   750.27    31.28k    87.37%
  10134223 requests in 1.00m, 1.26GB read
Requests/sec: 168623.70
Transfer/sec:     21.55MB

- Master -
Running 1m test @ http://0.0.0.0:18080/
  8 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   615.59us  262.44us  20.61ms   92.73%
    Req/Sec     1.64k   375.20     6.07k    68.39%
  783159 requests in 1.00m, 100.08MB read
Requests/sec:  13031.10
Transfer/sec:      1.67MB
```

These results show an improvement of 400% to 1000%. I couldn't believe the results so I ran the boost tests 3 more times, and got the same result.

if I'm not doing something horribly wrong (which I probably am) This PR is an extreme performance "boost" (Ha Ha) as well as an improvement in Crow's dependencies.