@mrozigor 2022-06-14
> Everything seems to be in order, PR seems to have no effect on compiler version and it doesn't seem to be any slower than the non-boost version.
> 
> @mrozigor Would you like to take a look or share some feedback? It's a pretty large PR so I would very much appreciate your opinion on it.

I've look into code and it looks pretty good. And great work to remove unnecessary dependency! ;)