@The-EDev 2022-06-15
I'll take one last look on the MacOS instructions and try to follow them to see if there are any problems.