@luca-schlecker 2022-06-15
There is something wrong with the `crow::utility::trim` function I have overlooked.
I'll fix it.