@luca-schlecker 2022-06-16
Fixed the `trim` usage and removed two redundant whitespace checks in `CookieParser`. Unit tests are still passing on my system.
This Pr should now be ready for merging. 🚀