@The-EDev 2022-06-17
There's an issue compiling on macos relating to query_string, I'm looking into it